[Append] Exclude Deleted Column (#676)

+++ clients/shared/append.go
@@ -21,7 +21,9 @@ func Append(dwh destination.DataWarehouse, tableData *optimization.TableData, op
 		return fmt.Errorf("failed to get table config: %w", err)
 	}
 
-	tableData.InMemoryColumns().DeleteColumn(constants.DeleteColumnMarker)
+	if opts.ShouldExcludeDeletedColumn {
+		tableData.InMemoryColumns().DeleteColumn(constants.DeleteColumnMarker)
+	}
 
 	// We don't care about srcKeysMissing because we don't drop columns when we append.
 	_, targetKeysMissing := columns.Diff(
